requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dnspython==2.6.1
fonttools==4.49.0
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
importlib_metadata==7.0.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pdf2image==1.17.0
pillow==10.2.0
protobuf==4.25.3
pyarrow==15.0.1
pydeck==0.8.1b0
Pygments==2.17.2
pymongo==4.6.2
pyparsing==3.1.2
pyperclip==1.8.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
streamlit==1.31.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
zipp==3.17.0
